[step4] sectionController를 lineController로 합치기
