+++ src/main/java/subway/controller/LineController.java
@@ -5,7 +5,10 @@ import org.springframework.web.bind.annotation.*;
 import subway.controller.dto.line.LineModifyRequest;
 import subway.controller.dto.line.LineResponse;
 import subway.controller.dto.line.LineSaveRequest;
+import subway.controller.dto.section.SectionResponse;
+import subway.controller.dto.section.SectionSaveRequest;
 import subway.service.LineCompositeService;
+import subway.service.SectionCompositeService;
 
 import java.net.URI;
 import java.util.List;
@@ -13,9 +16,11 @@ import java.util.List;
 @RestController
 public class LineController {
     private final LineCompositeService lineCompositeService;
+    private final SectionCompositeService sectionCompositeService;
 
-    public LineController(LineCompositeService lineCompositeService) {
+    public LineController(LineCompositeService lineCompositeService, SectionCompositeService sectionCompositeService) {
         this.lineCompositeService = lineCompositeService;
+        this.sectionCompositeService = sectionCompositeService;
     }
 
     @GetMapping(value = "/lines")
@@ -46,6 +51,25 @@ public class LineController {
     @DeleteMapping("/lines/{id}")
     public ResponseEntity<Void> deleteLine(@PathVariable Long id) {
         lineCompositeService.deleteLineById(id);
+        return ResponseEntity.noContent()
+                             .build();
+    }
+    @GetMapping(value = "/lines/{lineId}/sections")
+    public ResponseEntity<List<SectionResponse>> getSections(@PathVariable("lineId") Long lineId) {
+        return ResponseEntity.ok()
+                             .body(sectionCompositeService.findSectionsByLine(lineId));
+    }
+
+    @PostMapping("/lines/{lineId}/sections")
+    public ResponseEntity<SectionResponse> createSection(@PathVariable("lineId") Long lineId, @RequestBody SectionSaveRequest sectionSaveRequest) {
+        SectionResponse section = sectionCompositeService.saveSection(lineId, sectionSaveRequest);
+        return ResponseEntity.created(URI.create("/sections/" + section.getId()))
+                             .body(section);
+    }
+
+    @DeleteMapping("/lines/{lineId}/sections")
+    public ResponseEntity<Void> deleteSection(@PathVariable("lineId") Long lineId, @RequestParam("stationId") Long stationId) {
+        sectionCompositeService.deleteSectionByStationId(lineId, stationId);
         return ResponseEntity.noContent()
                              .build();
     }
